[/db_seeder] Refactor seeder script
