+++ scripts/db_seeder/db_seeder.py
@@ -1,0 +1,4 @@
+from populate_users import populate_users
+
+if __name__ == "__main__":
+    populate_users(25)
